requirements.txt
appdirs==1.4.4
attrdict==2.0.1
audioread==2.1.9
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
decorator==5.0.9
glcontext==2.3.4
greenlet==1.1.0
idna==2.10
imgui==1.3.0
joblib==1.0.1
librosa==0.8.1
llvmlite==0.36.0
moderngl==5.6.4
numba==0.53.1
numpy==1.20.3
packaging==20.9
Pillow==8.2.0
pooch==1.3.0
pycparser==2.20
pydux==0.2.2
PyOpenGL==3.1.5
pyparsing==2.4.7
PySDL2==0.9.7
PyYAML==5.4.1
requests==2.25.1
resampy==0.2.2
scikit-learn==0.24.2
scipy==1.6.3
simplejson==3.17.2
six==1.16.0
SoundFile==0.10.3.post1
SQLAlchemy==1.4.17
threadpoolctl==2.1.0
urllib3==1.26.5
yapf==0.31.0
